requirements.txt
flask==2.1.1

PyJWT~=2.3.0

Werkzeug==2.2.2